chore: split off orchestrator state into separate file
